separate out format readers
